refactor: 💡 extract parse date strings function to utils file
